requirements.txt
flet==0.27.1
loguru==0.7.3
python-dotenv==1.0.1
